pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_backend"

[project]
name = "dailybriefer"
version = "0.1.0"
description = "Local-only, objective morning news brief generator"
requires-python = ">=3.8"
dependencies = [
    "feedparser>=6.0.10",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src"]
